New params and validations for generator

+++ templates/Enterprise-Template/src/typescript/generator-botbuilder-enterprise/generators/app/index.js
@@ -1,12 +1,13 @@
 "use strict";
 const Generator = require("yeoman-generator");
 const chalk = require("chalk");
-const pkg = require("../../package.json");
 const path = require("path");
+const pkg = require("../../package.json");
 const _pick = require("lodash/pick");
 const _camelCase = require("lodash/camelCase");
 const _upperFirst = require("lodash/upperFirst");
 const _kebabCase = require("lodash/kebabCase");
+const fs = require("fs");
 
 const languages = [
   {
@@ -62,7 +63,8 @@ const bigBot =
 const tinyBot =
   " " + chalk.blue.bold("<") + " ** " + chalk.blue.bold(">") + " ";
 
-var botGenerationPath;
+var botGenerationPath = process.cwd();
+var isAlreadyCreated = false;
 
 module.exports = class extends Generator {
   constructor(args, opts) {
@@ -88,6 +90,13 @@ module.exports = class extends Generator {
       alias: "l"
     });
 
+    this.option("botGenerationPath", {
+      description: "The path where the bot will be generated",
+      type: String,
+      default: process.cwd(),
+      alias: "p"
+    });
+
     this.option("noPrompt", {
       description: "Do not prompt for any information or confirmation",
       type: Boolean,
@@ -96,8 +105,8 @@ module.exports = class extends Generator {
   }
 
   prompting() {
+    
     this.log(bigBot);
-
     // Validate language option
     if (
       this.options.botLang &&
@@ -120,7 +129,19 @@ module.exports = class extends Generator {
         type: "input",
         name: "botName",
         message: `What's the name of your bot?`,
-        default: this.options.botName ? this.options.botName : "enterprise-bot"
+        default: this.options.botName ? this.options.botName : "enterprise-bot",
+        validate: input => {
+          if (input.length > 12) {
+            this.log(
+              chalk.yellow(
+                "\n",
+                "WARNING: If the name of the bot has more than 12 characters, it could have some problems deploying some services"
+              )
+            );
+          }
+
+          return true;
+        }
       },
       {
         type: "input",
@@ -138,15 +159,46 @@ module.exports = class extends Generator {
         default: selectedLang
       },
       {
-        name: "finalConfirmation",
         type: "confirm",
+        name: "pathConfirmation",
+        message: "Do you want to change the location of the generation?",
+        default: false
+      },
+      {
+        when: function(response) {
+          return response.pathConfirmation === true;
+        },
+        type: "input",
+        name: "botGenerationPath",
+        message: "Where do you want to generate the bot?",
+        default: this.options.botGenerationPath
+          ? this.options.botGenerationPath
+          : process.cwd(),
+        validate: path => {
+          if (fs.existsSync(path)) {
+            return true;
+          }
+
+          this.log(
+            chalk.red("\n", "ERROR: This is not a valid path. Please try again")
+          );
+        },
+      },
+      {
+        type: "confirm",
+        name: "finalConfirmation",
         message: "Looking good. Shall I go ahead and create your new bot?",
         default: true
       }
     ];
 
     if (this.options.noPrompt) {
-      this.props = _pick(this.options, ["botName", "botDesc", "botLang"]);
+      this.props = _pick(this.options, [
+        "botName",
+        "botDesc",
+        "botLang",
+        "botGenerationPath"
+      ]);
 
       // Validate we have what we need, or we'll need to throw
       if (!this.props.botName) {
@@ -173,34 +225,50 @@ module.exports = class extends Generator {
     const templateName = "enterprise-bot";
     const botLang = this.props.botLang;
     const botDesc = this.props.botDesc;
-    const botName = _kebabCase(this.props.botName);
+    if (!this.props.botName.replace(/\s/g, "").length) {
+      this.props.botName = templateName;
+    }
+
+    const botName = _kebabCase(this.props.botName).replace(
+      /([^a-z0-9-]+)/gi,
+      ""
+    );
     const botNamePascalCase = _upperFirst(_camelCase(this.props.botName));
     const botNameCamelCase = _camelCase(this.props.botName);
+    botGenerationPath = path.join(botGenerationPath, botName);
+    if (this.props.botGenerationPath !== undefined) {
+      botGenerationPath = path.join(this.props.botGenerationPath, botName);
+    }
 
-    this.log("Current values:");
-    this.log(botName);
-    this.log(botNameCamelCase);
-    this.log(botDesc);
-    this.log(botLang);
+    if (fs.existsSync(botGenerationPath)){
+      isAlreadyCreated = true;
+      return;
+    }
+
+    this.log(chalk.magenta("\nCurrent values for the new bot:"));
+    this.log(chalk.magenta("Name: " + botName));
+    this.log(chalk.magenta("Description: " + botDesc));
+    this.log(chalk.magenta("Language: " + botLang));
+    this.log(chalk.magenta("Path: " + botGenerationPath + "\n"));
 
     this.fs.copy(
       this.templatePath(templateName, "cognitiveModels", "LUIS", botLang, "*"),
-      this.destinationPath(botNameCamelCase, "cognitiveModels", "LUIS")
+      this.destinationPath(botGenerationPath, "cognitiveModels", "LUIS")
     );
 
     this.fs.copy(
       this.templatePath(templateName, "cognitiveModels", "QnA", botLang, "*"),
-      this.destinationPath(botNameCamelCase, "cognitiveModels", "QnA")
+      this.destinationPath(botGenerationPath, "cognitiveModels", "QnA")
     );
 
     this.fs.copy(
       this.templatePath(templateName, "deploymentScripts", botLang, "*"),
-      this.destinationPath(botNameCamelCase, "deploymentScripts")
+      this.destinationPath(botGenerationPath, "deploymentScripts")
     );
 
     this.fs.copyTpl(
       this.templatePath(templateName, "_package.json"),
-      this.destinationPath(botNameCamelCase, "package.json"),
+      this.destinationPath(botGenerationPath, "package.json"),
       {
         name: botName,
         description: botDesc
@@ -209,7 +277,7 @@ module.exports = class extends Generator {
 
     this.fs.copyTpl(
       this.templatePath(templateName, "src", "_index.ts"),
-      this.destinationPath(botNameCamelCase, "src", "index.ts"),
+      this.destinationPath(botGenerationPath, "src", "index.ts"),
       {
         name: botName,
         description: botDesc,
@@ -220,7 +288,7 @@ module.exports = class extends Generator {
 
     this.fs.copyTpl(
       this.templatePath(templateName, "src", "_enterpriseBot.ts"),
-      this.destinationPath(botNameCamelCase, "src", `${botNameCamelCase}.ts`),
+      this.destinationPath(botGenerationPath, "src", `${botNameCamelCase}.ts`),
       {
         name: botName,
         description: botDesc,
@@ -231,7 +299,7 @@ module.exports = class extends Generator {
 
     this.fs.copy(
       this.templatePath(templateName, "src", "botServices.ts"),
-      this.destinationPath(botNameCamelCase, "src", "botServices.ts")
+      this.destinationPath(botGenerationPath, "src", "botServices.ts")
     );
 
     const commonFiles = [
@@ -246,7 +314,7 @@ module.exports = class extends Generator {
     commonFiles.forEach(fileName =>
       this.fs.copy(
         this.templatePath(templateName, fileName),
-        this.destinationPath(botNameCamelCase + "/" + fileName)
+        this.destinationPath(botGenerationPath, fileName)
       )
     );
 
@@ -261,17 +329,15 @@ module.exports = class extends Generator {
     commonDirectories.forEach(directory =>
       this.fs.copy(
         this.templatePath(templateName, "src", directory, "**", "*"),
-        this.destinationPath(botNameCamelCase, "src", directory)
+        this.destinationPath(botGenerationPath, "src", directory)
       )
     );
-
-    botGenerationPath = path.join(process.cwd(), botNameCamelCase);
   }
 
   install() {
-    if (this.props.finalConfirmation !== true) {
+    if (this.props.finalConfirmation !== true || isAlreadyCreated) {
       return;
-    }
+    }      
 
     process.chdir(botGenerationPath);
     this.installDependencies({ npm: true, bower: false });
@@ -279,14 +345,22 @@ module.exports = class extends Generator {
 
   end() {
     if (this.props.finalConfirmation === true) {
-      this.log(chalk.green("------------------------ "));
-      this.log(chalk.green(" Your new bot is ready!  "));
-      this.log(chalk.green("------------------------ "));
-      this.log(
-        "Open the " +
-          chalk.green.bold("README.md") +
-          " to learn how to run your bot. "
-      );
+      if(isAlreadyCreated){
+        this.log(chalk.red.bold("-------------------------------------------------------------------------------------------- "));
+        this.log(chalk.red.bold(" ERROR: It's seems like you already have a bot with the same name in the destination path. "));
+        this.log(chalk.red.bold(" Try again changing the name or the destination path or deleting the previous bot. "));
+        this.log(chalk.red.bold("-------------------------------------------------------------------------------------------- "));
+      }
+      else{
+        this.log(chalk.green("------------------------ "));
+        this.log(chalk.green(" Your new bot is ready!  "));
+        this.log(chalk.green("------------------------ "));
+        this.log(
+          "Open the " +
+            chalk.green.bold("README.md") +
+            " to learn how to run your bot. "
+        );
+      }
     } else {
       this.log(chalk.red.bold("-------------------------------- "));
       this.log(chalk.red.bold(" New bot creation was canceled. "));
